AdjustReadMe - Codacy
Per Codacy recommendation, Readme has been adjusted to include double spacing with the list.

An attempt to also ammend the null terminated strings issue made by codacy was attempted. an attempt to change strlen() to sizeof() resulted in errors (obviously) the dll would no longer inject into the target process. Perhaps somebody may comment on the future as to what exactly this error implies. . . I however do not see anybody crashing due to this as the null termination is accounted for when all strings are parsed anyways. In the end it will fall on the user for not double checking their document before running the application. Checks are in place for an empty config file and to me that is sufficient enough due to the nature of the product