CoreChanges (#1)
Summary
Basically I've made these changes to simplify the usage of the injection module. Previously users would need to have process name and absolute dll path ready to paste (or remembered, bless them) to be able to inject a module into the target process. This would take anywhere from 5seconds to a minute.

Now when a user wishes to inject a DLL into a process they just need to place the dll into the same directory as the executable and config file. Adjust the config file to declare the target process and target dll name. Launching the process will immediately start a process which will both search for and wait for the entered process to be running. Upon finding the target application an attempt to gain an open handle to the process is made. If successful the dll is injected into the process with LoadLibraryA

Changelog
- Automated Injection Process, information is now stored and loaded from a config file
- Error Management with smart error code output, handled via MessageBoxes
- Created a struct to hold the target information, a function is used to populate the struct which is then used in the initialization process.

* AdjustReadMe - Codacy